+++ test/tx-decorator/index.test.js
@@ -1,3 +1,3 @@
-describe('decorateTxParams', () => {
+test('decorateTxParams', () => {
     // @TODO test only src and not test builds
 });