+++ test/tx.test.js
@@ -118,9 +118,9 @@ describe('buy', () => {
 });
 
 // @TODO
-describe('create coin', () => {
-    const txItem = findTxItem(TX_TYPE.CREATE_COIN);
-});
+// describe('create coin', () => {
+//     const txItem = findTxItem(TX_TYPE.CREATE_COIN);
+// });
 
 describe('declare candidacy', () => {
     const txItem = findTxItem(TX_TYPE.DECLARE_CANDIDACY);
@@ -185,21 +185,21 @@ describe('unbond', () => {
 });
 
 // @TODO
-describe('set candidate on', () => {
-    const txItem = findTxItem(TX_TYPE.SET_CANDIDATE_ON);
-});
+// describe('set candidate on', () => {
+//     const txItem = findTxItem(TX_TYPE.SET_CANDIDATE_ON);
+// });
 
-describe('set candidate off', () => {
-    const txItem = findTxItem(TX_TYPE.SET_CANDIDATE_OFF);
-});
+// describe('set candidate off', () => {
+//     const txItem = findTxItem(TX_TYPE.SET_CANDIDATE_OFF);
+// });
 
-describe('create multisig', () => {
-    const txItem = findTxItem(TX_TYPE.CREATE_MULTISIG);
-});
+// describe('create multisig', () => {
+//     const txItem = findTxItem(TX_TYPE.CREATE_MULTISIG);
+// });
 
-describe('multisend', () => {
-    const txItem = findTxItem(TX_TYPE.MULTISEND);
-});
+// describe('multisend', () => {
+//     const txItem = findTxItem(TX_TYPE.MULTISEND);
+// });
 
 describe('redeem check', () => {
     const txItem = findTxItem(TX_TYPE.REDEEM_CHECK);