+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - restore support of `MoveStake` tx data
 - add `disableValidation` option to tx data constructors
 - add `disableDecorationParams` and `disableValidation` options to `prepareTx`
-- fix `estimateTxCommission` to work with empty data params in `loose: false` mode
+- fix `estimateTxCommission` to work with empty data params and coin symbol params in `loose: false` mode
 - **BREAKING** add fields to VoteCommission tx data
 - **DEPRECATE** `direct` option in `estimateTxCommission`, use `loose` option with reversed value instead